INCOMPLETE: make shader_t and rlsl api agnostic
