+++ include/rocket/shader.hpp
@@ -3,14 +3,15 @@
 
 #include "types.hpp"
 #include <filesystem>
-#include <rocket/rgl.hpp>
 #include <string>
+#include <vector>
 
 namespace rocket {
     enum class shader_id_t;
 }
 
 namespace rgl {
+    typedef unsigned int shader_program_t;
     rgl::shader_program_t get_shader(rocket::shader_id_t shid);
 }
 
@@ -19,52 +20,101 @@ namespace rocket {
         vert_frag,
     };
 
-    class shader_t {
-    private:
-        unsigned int glshaderv = rGL_SHADER_INVALID;
-        unsigned int glshaderf = rGL_SHADER_INVALID;
-        unsigned int glprogram = rGL_SHADER_INVALID;
+    class shader_i {
+    protected:
         shader_type type;
-        std::string vcode = "";
-        std::string fcode = "";
-
         std::string name = "NON_RLSL_SHADER";
         std::string rlsl_version = "NOT_COMPILED_BY_RLSL";
 
-        unsigned int vao { rGL_SHADER_INVALID },
-                     vbo { rGL_SHADER_INVALID };
+        std::string vcode;
+        std::string fcode;
         
         friend class renderer_2d_i;
         friend class opengl_renderer_2d;
         friend class vulkan_renderer_2d;
         friend class renderer_3d;
         friend rgl::shader_program_t get_shader(shader_id_t shid);
-        friend shader_t load_from_rlsl_source(shader_type type, std::string rlsl);
         friend void shader_provider_reset();
-    private:
-        void shader_init();
-        void parse(const std::vector<std::string> &lines, std::filesystem::path shader_workingdir);
+    protected:
+        virtual void shader_init() = 0;
+        virtual void parse(const std::vector<std::string> &lines, std::filesystem::path shader_workingdir) = 0;
     public:
-        void set_uniform(std::string name, float value);
-        void set_uniform(std::string name, int value);
-        void set_uniform(std::string name, vec2f_t value);
-        void set_uniform(std::string name, vec3f_t value);
-        void set_uniform(std::string name, vec4f_t value);
-        void set_uniform(std::string name, mat4 value);
-        void set_uniform_raw(std::string name, unsigned int type, const void* data, int count);
+        virtual void set_parameter(std::string name, float value) = 0;
+        virtual void set_parameter(std::string name, int value) = 0;
+        virtual void set_parameter(std::string name, vec2f_t value) = 0;
+        virtual void set_parameter(std::string name, vec3f_t value) = 0;
+        virtual void set_parameter(std::string name, vec4f_t value) = 0;
+        virtual void set_parameter(std::string name, mat4 value) = 0;
+        virtual void set_parameter_raw(std::string name, unsigned int type, const void* data, int count) = 0;
     public:
-        bool operator==(const shader_t &other) const {
-            return glprogram == other.glprogram;
-        }
+        virtual bool operator==(const shader_i &other) const = 0;
     public:
-        shader_t(shader_type type, std::string vcode, std::string fcode, std::string name = "NON_RLSL_SHADER");
+        shader_i(shader_type type, std::string vcode, std::string fcode, std::string name = "NON_RLSL_SHADER");
         /// @brief Loads a shader from a file (.rlsl)
-        shader_t(shader_type type, std::string rlsl_shader_path);
-        shader_t();
-        static shader_t load_from_rlsl_source(shader_type type, std::string rlsl);
-        static shader_t rectangle(rgba_color fill_color);
+        shader_i(shader_type type, std::filesystem::path rlsl_shader_path);
+        shader_i(shader_type type, const std::string &rlsl_source);
+        shader_i();
     public:
-        ~shader_t();
+        virtual ~shader_i();
+    };
+
+    class opengl_shader_t : public shader_i {
+    private:
+        uint32_t glshaderv = 0xDEADBEEFU;
+        uint32_t glshaderf = 0xDEADBEEFU;
+        uint32_t glprogram = 0xDEADBEEFU;
+
+        uint32_t vao = 0xDEADBEEFU;
+        uint32_t vbo = 0xDEADBEEFU;
+
+        friend class opengl_renderer_2d;
+    private:
+        void shader_init() override;
+        void parse(const std::vector<std::string> &lines, std::filesystem::path shader_workingdir) override;
+    public:
+        void set_parameter(std::string name, float value) override;
+        void set_parameter(std::string name, int value) override;
+        void set_parameter(std::string name, vec2f_t value) override;
+        void set_parameter(std::string name, vec3f_t value) override;
+        void set_parameter(std::string name, vec4f_t value) override;
+        void set_parameter(std::string name, mat4 value) override;
+        void set_parameter_raw(std::string name, unsigned int type, const void* data, int count) override;
+    public:
+        bool operator==(const shader_i &other) const override;
+    public:
+        opengl_shader_t(shader_type type, std::string vcode, std::string fcode, std::string name = "NON_RLSL_SHADER");
+        /// @brief Loads a shader from a file (.rlsl)
+        opengl_shader_t(shader_type type, std::filesystem::path rlsl_shader_path);
+        opengl_shader_t(shader_type type, const std::string &rlsl_source);
+        opengl_shader_t();
+    public:
+        ~opengl_shader_t() override;
+    };
+
+    struct vulkan_shader_t : public shader_i {
+    private:
+        friend class vulkan_renderer_2d;
+    private:
+        void shader_init() override;
+        void parse(const std::vector<std::string> &lines, std::filesystem::path shader_workingdir) override;
+    public:
+        void set_parameter(std::string name, float value) override;
+        void set_parameter(std::string name, int value) override;
+        void set_parameter(std::string name, vec2f_t value) override;
+        void set_parameter(std::string name, vec3f_t value) override;
+        void set_parameter(std::string name, vec4f_t value) override;
+        void set_parameter(std::string name, mat4 value) override;
+        void set_parameter_raw(std::string name, unsigned int type, const void* data, int count) override;
+    public:
+        bool operator==(const shader_i &other) const override;
+    public:
+        vulkan_shader_t(shader_type type, std::string vcode, std::string fcode, std::string name = "NON_RLSL_SHADER");
+        /// @brief Loads a shader from a file (.rlsl)
+        vulkan_shader_t(shader_type type, std::filesystem::path rlsl_shader_path);
+        vulkan_shader_t(shader_type type, const std::string &rlsl_source);
+        vulkan_shader_t();
+    public:
+        ~vulkan_shader_t() override;
     };
 }
