+++ src/rocket/gfx/renderer/vulkan_renderer.cpp
@@ -76,7 +76,7 @@ namespace rocket {
     void vulkan_renderer_2d::draw_text(const rocket::text_t& text_, rocket::vec2f_t position) {
     }
 
-    void vulkan_renderer_2d::draw_shader(shader_t shader) {
+    void vulkan_renderer_2d::draw_shader(const shader_i &shader) {
     }
 
     void vulkan_renderer_2d::set_wireframe(bool x) {
